requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
PyJWT==2.3.0
bcrypt==3.2.0
PyPDF2==3.0.1
openai==0.27.0
python-dotenv==0.20.0